requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio

# Segmentation Models
segmentation-models-pytorch==0.3.3
albumentations==1.3.1

# WSI Processing
openslide-python==1.3.0
opencv-python==4.8.0.74
scikit-image==0.21.0
Pillow==10.0.0

# Experiment Tracking
wandb==0.15.12

# Metrics & Visualization
torchmetrics==1.2.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
PyYAML==6.0.1
h5py==3.9.0

# Development Tools
ipython==8.14.0
jupyter==1.0.0
tensorboard==2.14.0

# Note: Use setup_environment.sh for complete installation with CUDA support
# For manual pip install: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118